added gate directory validation at game load
