+++ src/main/java/com/TravelGatesMod/TravelGates/GUI/DestinationSelectionScreen.java
@@ -1,36 +1,32 @@
-package com.example.TravelGates.GUI;
+package com.TravelGatesMod.TravelGates.GUI;
 
-import com.example.TravelGates.travelgates;
-import com.example.TravelGates.util.GateInfo;
-import com.example.TravelGates.util.GateInfoHandler;
+import com.TravelGatesMod.TravelGates.travelgates;
+import com.TravelGatesMod.TravelGates.util.GateInfo;
+import com.TravelGatesMod.TravelGates.util.GateInfoHandler;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.button.Button;
-import net.minecraft.client.gui.widget.button.CheckboxButton;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.StringTextComponent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.Dictionary;
 import java.util.ListIterator;
 
-public class GateWhiteListScreen extends CheckedItemScreen {
+public class DestinationSelectionScreen extends Screen {
+
 
     public static final int WIDTH = 179;
     public static final int HEIGHT = 151;
     public static int PageNum = 0;
     private GateScreen PARENTSCREEN;
-    boolean dumbBool;
 
     private static final Logger LOGGER = LogManager.getLogger();
 
     private ResourceLocation GUI = new ResourceLocation(travelgates.MOD_ID, "textures/gui/destination_select_gui.png");
 
-    //public Dictionary<String, Boolean>;
-
-    protected GateWhiteListScreen(GateScreen screen) {
-        super(new StringTextComponent("Select WhiteList"));
+    protected DestinationSelectionScreen(GateScreen screen) {
+        super(new StringTextComponent("Select Gate Destination"));
 
         PARENTSCREEN = screen;
     }
@@ -55,12 +51,8 @@ public class GateWhiteListScreen extends CheckedItemScreen {
         {
             info = (GateInfo)iterator.next();
             String ID = info.GATE_ID;
-            if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(info.GATE_ID))
-            {
-                dumbBool = true;
-            }
-            addButton(new GateCheckboxButton(x + 10, (y + (10)+ (i*27)),160, 20, info.GATE_ID, dumbBool,this));
-            dumbBool = false;
+
+            addButton(new Button(x + 10, (y + (10)+ (i*27)),160, 20, info.GATE_ID, button -> SetDestination(ID)));
         }
 
 
@@ -74,56 +66,31 @@ public class GateWhiteListScreen extends CheckedItemScreen {
             addButton(new Button(x+10 , (y + 125),30, 20, "Back", button -> PreviousPage()));
         }
 
-        addButton(new Button(x+65 , (y + 125),50, 20, "Accept", button -> Accept()));
+        addButton(new Button(x+65 , (y + 125),50, 20, "Cancel", button -> Cancel()));
     }
 
-    @Override
-    public void AddItemToList(String ID)
+    public void SetDestination(String ID)
     {
-        if(!this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(ID))
-        {
-            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.add(ID);
-        }
-        else
-        {
-            LOGGER.warn("ID was found in whitelist when it should not have been there");
-        }
-
+        PARENTSCREEN.CallingGateInfo.DESTINATION_GATE_ID = ID;
+        PARENTSCREEN.open();
     }
 
-    @Override
-    public void RemoveItemFromList(String ID)
-    {
-        if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(ID))
-        {
-            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.remove(ID);
-        }
-        else
-        {
-            LOGGER.warn("ID was not found in whitelist when it should have been");
-        }
-    }
-
-    @Override
     public void NextPage()
     {
         PageNum++;
         this.open(PARENTSCREEN);
     }
 
-    @Override
     public void PreviousPage()
     {
         PageNum--;
         this.open(PARENTSCREEN);
     }
 
-    @Override
-    public void Accept()
+    public void Cancel()
     {
         PARENTSCREEN.open();
     }
-
 
     @Override
     public void render(int mouseX, int mouseY, float partialTicks)
@@ -135,12 +102,16 @@ public class GateWhiteListScreen extends CheckedItemScreen {
         this.blit(relX,relY,0,0,WIDTH,HEIGHT);
         super.render(mouseX,mouseY,partialTicks);
 
+        //this.drawString(this.font, I18n.format("selectWorld.enterName"), relX, 47, -6250336);
+        //this.GateIDField.render(mouseX, mouseY, partialTicks);
+
+        //this.GateIDField.active = true;
 
     }
 
     public static void open(GateScreen parentScreen)
     {
-        Minecraft.getInstance().displayGuiScreen(new GateWhiteListScreen(parentScreen));
+        Minecraft.getInstance().displayGuiScreen(new DestinationSelectionScreen(parentScreen));
 
     }
 
@@ -149,4 +120,5 @@ public class GateWhiteListScreen extends CheckedItemScreen {
     {
         return false;
     }
+
 }
