+++ src/main/java/com/TravelGatesMod/TravelGates/GUI/GateBlackListScreen.java
@@ -1,10 +1,9 @@
-package com.example.TravelGates.GUI;
+package com.TravelGatesMod.TravelGates.GUI;
 
-import com.example.TravelGates.travelgates;
-import com.example.TravelGates.util.GateInfo;
-import com.example.TravelGates.util.GateInfoHandler;
+import com.TravelGatesMod.TravelGates.travelgates;
+import com.TravelGatesMod.TravelGates.util.GateInfo;
+import com.TravelGatesMod.TravelGates.util.GateInfoHandler;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.button.Button;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.StringTextComponent;
@@ -12,22 +11,22 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ListIterator;
-import java.util.Optional;
 
-public class DestinationSelectionScreen extends Screen {
-
+public class GateBlackListScreen extends CheckedItemScreen {
 
     public static final int WIDTH = 179;
     public static final int HEIGHT = 151;
     public static int PageNum = 0;
     private GateScreen PARENTSCREEN;
+    boolean dumbBool;
 
     private static final Logger LOGGER = LogManager.getLogger();
 
     private ResourceLocation GUI = new ResourceLocation(travelgates.MOD_ID, "textures/gui/destination_select_gui.png");
 
-    protected DestinationSelectionScreen(GateScreen screen) {
-        super(new StringTextComponent("Select Gate Destination"));
+
+    protected GateBlackListScreen(GateScreen screen) {
+        super(new StringTextComponent("Select BlackList"));
 
         PARENTSCREEN = screen;
     }
@@ -52,8 +51,13 @@ public class DestinationSelectionScreen extends Screen {
         {
             info = (GateInfo)iterator.next();
             String ID = info.GATE_ID;
-
-            addButton(new Button(x + 10, (y + (10)+ (i*27)),160, 20, info.GATE_ID, button -> SetDestination(ID)));
+            if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(info.GATE_ID))
+            {
+                LOGGER.debug("BlackList comparison true");
+                dumbBool = true;
+            }
+            addButton(new GateCheckboxButton(x + 10, (y + (10)+ (i*27)),160, 20, info.GATE_ID, dumbBool,this));
+            dumbBool = false;
         }
 
 
@@ -67,31 +71,56 @@ public class DestinationSelectionScreen extends Screen {
             addButton(new Button(x+10 , (y + 125),30, 20, "Back", button -> PreviousPage()));
         }
 
-        addButton(new Button(x+65 , (y + 125),50, 20, "Cancel", button -> Cancel()));
+        addButton(new Button(x+65 , (y + 125),50, 20, "Accept", button -> Accept()));
     }
 
-    public void SetDestination(String ID)
+    @Override
+    public void AddItemToList(String ID)
     {
-        PARENTSCREEN.CallingGateInfo.DESTINATION_GATE_ID = ID;
-        PARENTSCREEN.open();
+        if(!this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(ID))
+        {
+            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.add(ID);
+        }
+        else
+        {
+            LOGGER.warn("ID was found in BlackList when it should not have been there");
+        }
+
     }
 
+    @Override
+    public void RemoveItemFromList(String ID)
+    {
+        if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(ID))
+        {
+            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.remove(ID);
+        }
+        else
+        {
+            LOGGER.warn("ID was not found in BlackList when it should have been");
+        }
+    }
+
+    @Override
     public void NextPage()
     {
         PageNum++;
         this.open(PARENTSCREEN);
     }
 
+    @Override
     public void PreviousPage()
     {
         PageNum--;
         this.open(PARENTSCREEN);
     }
 
-    public void Cancel()
+    @Override
+    public void Accept()
     {
         PARENTSCREEN.open();
     }
+
 
     @Override
     public void render(int mouseX, int mouseY, float partialTicks)
@@ -103,16 +132,12 @@ public class DestinationSelectionScreen extends Screen {
         this.blit(relX,relY,0,0,WIDTH,HEIGHT);
         super.render(mouseX,mouseY,partialTicks);
 
-        //this.drawString(this.font, I18n.format("selectWorld.enterName"), relX, 47, -6250336);
-        //this.GateIDField.render(mouseX, mouseY, partialTicks);
-
-        //this.GateIDField.active = true;
 
     }
 
     public static void open(GateScreen parentScreen)
     {
-        Minecraft.getInstance().displayGuiScreen(new DestinationSelectionScreen(parentScreen));
+        Minecraft.getInstance().displayGuiScreen(new GateBlackListScreen(parentScreen));
 
     }
 
@@ -121,5 +146,4 @@ public class DestinationSelectionScreen extends Screen {
     {
         return false;
     }
-
 }
