+++ src/main/java/com/TravelGatesMod/TravelGates/GUI/GateSelectionList.java
@@ -1,0 +1,10 @@
+package com.TravelGatesMod.TravelGates.GUI;
+
+import net.minecraft.client.gui.screen.ReadBookScreen;
+
+public class GateSelectionList extends ReadBookScreen {
+
+
+
+
+}
