+++ src/main/java/com/TravelGatesMod/TravelGates/GUI/GateWhiteListScreen.java
@@ -1,19 +1,18 @@
-package com.example.TravelGates.GUI;
+package com.TravelGatesMod.TravelGates.GUI;
 
-import com.example.TravelGates.travelgates;
-import com.example.TravelGates.util.GateInfo;
-import com.example.TravelGates.util.GateInfoHandler;
+import com.TravelGatesMod.TravelGates.travelgates;
+import com.TravelGatesMod.TravelGates.util.GateInfo;
+import com.TravelGatesMod.TravelGates.util.GateInfoHandler;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.widget.button.Button;
 import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.StringTextComponent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ListIterator;
 
-public class GateBlackListScreen extends CheckedItemScreen {
+public class GateWhiteListScreen extends CheckedItemScreen {
 
     public static final int WIDTH = 179;
     public static final int HEIGHT = 151;
@@ -25,9 +24,10 @@ public class GateBlackListScreen extends CheckedItemScreen {
 
     private ResourceLocation GUI = new ResourceLocation(travelgates.MOD_ID, "textures/gui/destination_select_gui.png");
 
+    //public Dictionary<String, Boolean>;
 
-    protected GateBlackListScreen(GateScreen screen) {
-        super(new StringTextComponent("Select BlackList"));
+    protected GateWhiteListScreen(GateScreen screen) {
+        super(new StringTextComponent("Select WhiteList"));
 
         PARENTSCREEN = screen;
     }
@@ -52,9 +52,8 @@ public class GateBlackListScreen extends CheckedItemScreen {
         {
             info = (GateInfo)iterator.next();
             String ID = info.GATE_ID;
-            if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(info.GATE_ID))
+            if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(info.GATE_ID))
             {
-                LOGGER.debug("BlackList comparison true");
                 dumbBool = true;
             }
             addButton(new GateCheckboxButton(x + 10, (y + (10)+ (i*27)),160, 20, info.GATE_ID, dumbBool,this));
@@ -78,13 +77,13 @@ public class GateBlackListScreen extends CheckedItemScreen {
     @Override
     public void AddItemToList(String ID)
     {
-        if(!this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(ID))
+        if(!this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(ID))
         {
-            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.add(ID);
+            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.add(ID);
         }
         else
         {
-            LOGGER.warn("ID was found in BlackList when it should not have been there");
+            LOGGER.warn("ID was found in whitelist when it should not have been there");
         }
 
     }
@@ -92,13 +91,13 @@ public class GateBlackListScreen extends CheckedItemScreen {
     @Override
     public void RemoveItemFromList(String ID)
     {
-        if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.contains(ID))
+        if(this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.contains(ID))
         {
-            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_BLACKLIST.remove(ID);
+            this.PARENTSCREEN.CallingGateInfo.ARRIVAL_WHITELIST.remove(ID);
         }
         else
         {
-            LOGGER.warn("ID was not found in BlackList when it should have been");
+            LOGGER.warn("ID was not found in whitelist when it should have been");
         }
     }
 
@@ -138,7 +137,7 @@ public class GateBlackListScreen extends CheckedItemScreen {
 
     public static void open(GateScreen parentScreen)
     {
-        Minecraft.getInstance().displayGuiScreen(new GateBlackListScreen(parentScreen));
+        Minecraft.getInstance().displayGuiScreen(new GateWhiteListScreen(parentScreen));
 
     }
