+++ src/main/java/com/TravelGatesMod/TravelGates/blocks/QuickGateItem.java
@@ -1,10 +1,9 @@
-package com.example.TravelGates.blocks;
+package com.TravelGatesMod.TravelGates.blocks;
 
-import com.example.TravelGates.travelgates;
+import com.TravelGatesMod.TravelGates.travelgates;
 import net.minecraft.block.Block;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
 
 public class QuickGateItem extends BlockItem {
     public QuickGateItem(Block blockIn) {
