--- src/main/java/com/example/TravelGates/GUI/GateSelectionList.java
@@ -1,24 +1,0 @@
-package com.example.TravelGates.GUI;
-
-import com.example.TravelGates.util.GateInfo;
-import com.example.TravelGates.util.GateInfoHandler;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.IGuiEventListener;
-import net.minecraft.client.gui.screen.EditBookScreen;
-import net.minecraft.client.gui.screen.ReadBookScreen;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
-import net.minecraft.util.IItemProvider;
-
-import javax.annotation.Nullable;
-import java.util.AbstractList;
-import java.util.Optional;
-
-public class GateSelectionList extends ReadBookScreen {
-
-
-
-
-}
